+++ chess/board.py
@@ -51,8 +51,18 @@ class Board:                                       #se crea un tablero de ajedre
             board_str += "\n"
         return board_str
     
-    def get_piece(self, row, col):             #devuelve la pieza que se encuentra en una posición específica del tablero. Toma como parámetros la fila (row) y la columna (col) y devuelve la pieza ubicada en esa celda, o None si la celda está vacía.
-            return self.__positions__ [row] [col]
-    
+    def get_piece(self, row, col):
+        if not (
+            0 <= row < 8 or 0 <= col < 8
+        ):
+            raise OutOfBoard()
+        return self.__positions__[row][col]
+
     def set_piece(self, row, col, piece):      #mueve una pieza en una posición específica del tablero al actualizar las posiciones
         self.__positions__[row][col] = piece
+
+
+    def move(self, from_row, from_col, to_row, to_col):
+        origin = self.get_piece(from_row, from_col)
+        self.set_piece(to_row, to_col, origin)
+        self.set_piece(from_row, from_col, None)