+++ chess/pawn.py
@@ -1,4 +1,5 @@
-from pieces import Piece      #PEON CLASE
+from chess.pieces import Piece      #PEON CLASE
 
 class Pawn(Piece):
-    ...
+    white_str = "♜"
+    black_str = "♖"