+++ chess/pieces.py
@@ -9,3 +9,6 @@ class Piece:                       #HERENCIA PIEZAS
             return self.white_str
         else:
             return self.black_str
+        
+    def get_color(self):
+        return self.__color__